backend/requirements.txt
# ===============================================
# LeadLab Backend - Python Dependencies
# ===============================================
# Python 3.11+ required

# FastAPI and Web Framework
fastapi>=0.109.0,<0.120.0
uvicorn[standard]>=0.24.0,<0.30.0
python-multipart>=0.0.5
pydantic>=2.0.0,<3.0.0
pydantic-settings>=2.0.0
email-validator>=2.0.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
PyJWT==2.8.0
bcrypt==4.0.1
cryptography==41.0.7
python-dotenv==1.0.0

# Database
sqlalchemy>=2.0.23,<3.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.0
pymysql>=1.1.0

# Payment Processing
stripe==8.4.0

# ML and Data Analysis
scikit-learn>=1.4.0
pandas>=2.2.0
numpy>=1.26.0
joblib==1.3.2
# lightgbm==4.1.0  # CMake compatibility issue - uncomment if needed
category-encoders==2.6.3
feature-engine==1.6.2
optuna==3.4.0

# HTTP Clients
httpx>=0.25.1
requests==2.31.0
aiohttp==3.9.1

# Web Scraping
beautifulsoup4==4.12.2
lxml==4.9.3

# Cache, Performance & Monitoring
redis==5.0.1
prometheus-client>=0.19.0
sentry-sdk==1.35.0
psutil>=5.9.5

# Email & Communication
aiosmtplib>=2.0.0
jinja2==3.1.2
weasyprint==60.2
pdfkit==1.0.0
icalendar==6.3.1

# Date & Time Utilities
pytz==2025.2
python-dateutil==2.9.0

# Development & Testing Tools
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.0
isort==5.12.0
